add prettier and apply it

+++ contracts/test/PermitTest.sol
@@ -9,7 +9,6 @@ import "@openzeppelin/contracts/utils/introspection/ERC165.sol";
 import "@openzeppelin/contracts/token/ERC1155/IERC1155.sol";
 
 contract PermitTest is IERC1155Receiver, ERC165 {
-
     IMaidCoin private maidCoin;
     IMaid private maid;
     INursePart private nursePart;
@@ -25,61 +24,37 @@ contract PermitTest is IERC1155Receiver, ERC165 {
     }
 
     function maidCoinPermitTest(
-        uint amount,
-        uint deadline,
-        uint8 v, bytes32 r, bytes32 s
+        uint256 amount,
+        uint256 deadline,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
     ) external {
-        maidCoin.permit(
-            msg.sender,
-            address(this),
-            amount,
-            deadline,
-            v, r, s
-        );
-        maidCoin.transferFrom(
-            msg.sender,
-            address(this),
-            amount
-        );
+        maidCoin.permit(msg.sender, address(this), amount, deadline, v, r, s);
+        maidCoin.transferFrom(msg.sender, address(this), amount);
     }
 
     function maidPermitTest(
-        uint id,
-        uint deadline,
-        uint8 v, bytes32 r, bytes32 s
+        uint256 id,
+        uint256 deadline,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
     ) external {
-        maid.permit(
-            address(this),
-            id,
-            deadline,
-            v, r, s
-        );
-        maid.transferFrom(
-            msg.sender,
-            address(this),
-            id
-        );
+        maid.permit(address(this), id, deadline, v, r, s);
+        maid.transferFrom(msg.sender, address(this), id);
     }
 
     function nursePartPermitTest(
-        uint id,
-        uint amount,
-        uint deadline,
-        uint8 v, bytes32 r, bytes32 s
+        uint256 id,
+        uint256 amount,
+        uint256 deadline,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
     ) external {
-        nursePart.permit(
-            msg.sender,
-            address(this),
-            deadline,
-            v, r, s
-        );
-        nursePart.safeTransferFrom(
-            msg.sender,
-            address(this),
-            id,
-            amount,
-            ""
-        );
+        nursePart.permit(msg.sender, address(this), deadline, v, r, s);
+        nursePart.safeTransferFrom(msg.sender, address(this), id, amount, "");
     }
 
     function onERC1155Received(
@@ -103,8 +78,6 @@ contract PermitTest is IERC1155Receiver, ERC165 {
     }
 
     function supportsInterface(bytes4 interfaceId) public view virtual override(ERC165, IERC165) returns (bool) {
-        return
-            interfaceId == type(IERC1155).interfaceId ||
-            super.supportsInterface(interfaceId);
+        return interfaceId == type(IERC1155).interfaceId || super.supportsInterface(interfaceId);
     }
 }
